Move FileDiff::Change(s) into VCS namespace
Move FileDiff::Change and FileDiff::Changes subclasses into the VCS
namespace. All repository-related logic should live under the VCS
namespace.
