+++ spec/integrations/vcs/file_diff/changes/movement_spec.rb
@@ -1,29 +1,30 @@
 # frozen_string_literal: true
 
-RSpec.describe FileDiff::Changes::Addition, type: :model do
-  subject             { revision }
+RSpec.describe VCS::FileDiff::Changes::Movement, type: :model do
+  subject             { commit }
   let(:change)        { file_diffs.find(diff.id).changes.first }
   let(:parent_change) { file_diffs.find(parent_diff.id).changes.first }
-  let(:parent)        { create :file_resource, name: 'p-name' }
-  let(:file)          { create :file_resource, name: 'c-name', parent: parent }
-  let(:revision)      { create :revision }
-  let(:file_diffs)    { revision.file_diffs.reload }
+  let(:parent)        { create :vcs_staged_file, name: 'p-name' }
+  let(:file)          { create :vcs_staged_file, name: 'c-name' }
+  let(:commit)        { create :vcs_commit }
+  let(:file_diffs)    { commit.file_diffs.reload }
+
+  before { file.update(parent: parent) }
+
   let!(:diff) do
-    create :file_diff,
-           file_resource: file,
-           current_snapshot: file.current_snapshot,
-           previous_snapshot: nil,
-           revision: revision
+    create :vcs_file_diff,
+           current_snapshot_id: file.current_snapshot_id,
+           previous_snapshot_id: file.current_snapshot_id_before_last_save,
+           commit: commit
   end
   let!(:parent_diff) do
-    create :file_diff,
-           file_resource: parent,
+    create :vcs_file_diff,
            current_snapshot: parent.current_snapshot,
            previous_snapshot: nil,
-           revision: revision
+           commit: commit
   end
 
-  before { revision.assign_attributes(title: 'origin', is_published: true) }
+  before { commit.assign_attributes(title: 'origin', is_published: true) }
 
   describe 'validation: must_not_unselect_addition_of_parent' do
     let(:hook)  { nil }
@@ -33,7 +34,7 @@ RSpec.describe FileDiff::Changes::Addition, type: :model do
     it 'adds an error' do
       is_expected.to be_invalid
       expect(change.errors.full_messages).to eq [
-        "You cannot add 'c-name' without adding its parent folder 'p-name'"
+        "You cannot move 'c-name' without adding its parent folder 'p-name'"
       ]
     end
