+++ spec/models/vcs/file_diff/changes/deletion_spec.rb
@@ -2,9 +2,9 @@
 
 require 'models/shared_examples/being_a_file_diff_change.rb'
 
-RSpec.describe FileDiff::Changes::Deletion, type: :model do
+RSpec.describe VCS::FileDiff::Changes::Deletion, type: :model do
   subject(:change)  { described_class.new(diff: diff) }
-  let(:diff)        { instance_double FileDiff }
+  let(:diff)        { instance_double VCS::FileDiff }
 
   it_should_behave_like 'being a file diff change' do
     before { allow(diff).to receive(:ancestor_path).and_return 'path' }
