+++ spec/models/vcs/file_diff/changes/rename_spec.rb
@@ -2,9 +2,9 @@
 
 require 'models/shared_examples/being_a_file_diff_change.rb'
 
-RSpec.describe FileDiff::Changes::Rename, type: :model do
+RSpec.describe VCS::FileDiff::Changes::Rename, type: :model do
   subject(:change)  { described_class.new(diff: diff) }
-  let(:diff)        { instance_double FileDiff }
+  let(:diff)        { instance_double VCS::FileDiff }
 
   it_should_behave_like 'being a file diff change' do
     before { allow(diff).to receive(:ancestor_path).and_return 'path' }
@@ -18,8 +18,8 @@ RSpec.describe FileDiff::Changes::Rename, type: :model do
 
   describe '#unapply' do
     subject { change.current_snapshot }
-    let(:current_snapshot) { instance_double FileResource::Snapshot }
-    let(:previous_snapshot) { instance_double FileResource::Snapshot }
+    let(:current_snapshot) { instance_double VCS::FileSnapshot }
+    let(:previous_snapshot) { instance_double VCS::FileSnapshot }
 
     before do
       allow(change).to receive(:current_snapshot).and_return current_snapshot
